created host and web extension files moved common code around and fixed up import paths
