+++ yara/src/common/completionProvider.ts
