+++ yara/src/common/configuration.ts
