+++ yara/src/common/definitionProvider.ts
