+++ yara/src/common/helpers.ts
