+++ yara/src/common/hoverProvider.ts
