+++ yara/src/common/referenceProvider.ts
