+++ yara/src/common/snippetProvider.ts
